feat(Detail DetImages): morular image into Detail
